extract util
